refactor(components/navigation): change file naming convention
