+++ app/components/navigation/index.ts
@@ -1,4 +1,4 @@
-export * from './bar'
-export * from './list-item'
+export * from './navigation-bar'
+export * from './navigation-list-item'
 export * from './sidebar'
 export * from './sidebar-section-item'
